Add PMMR backend buffer, make PMMR Backend mutable
Continuing on implementing the PMMR persistent backend, this
introduces a couple major improvements:

* The Backend trait now has &mut self methods, and an &mut
reference in PMMR. This simplifies the implementation of all
backends by not forcing them to be interior mutable. Slight
drawback is that a backend can't be used directly as long as it's
used by a PMMR instance.
* Introduced a buffer in the PMMR persistent backend to allow
reads before the underlying files are fully flushed. Implemented
with a temporary VecBackend.

Also some minor bug fixes and documentation improvements. More
tests.

Next step: implement main data file compaction, flushing the
remove log and adding to a skip list instead.