Integrate PMMR and its persistent backend with the Chain
Wrap sum tree manipulations in a unit of work that can easily
abandoned. Use it in the chain pipeline to apply latest block
on top of the current state and check the new tree roots.
Update BlockHeader to have the root hash of the 3 sum trees.

This is expected to break a lot of things, more commits to fix
that up to follow. Missing parts are making sure the Backend
has no side effects until sync (there are a couple holes) and
the handling of fork additions and switching, which is going
to be tricky.