Handling of forks by rewinding the state
Implemented a new rewind method on the PMMR and persistent sum
tree implementations to be able to rollback some part of the
state to handle forks on past blocks. The chain pipeline
identifies the common block and rewinds to that state before
calculating state roots.
Also added a positional index for range proofs to help rewind
(and potentially future pruning).

Added more tests on branching (still more required). Now need to
integrate with modules sitting above the chain.