Fix `DuplicateCommitment` check
We need to check whether the commitment is actually in the MMR or
not as the index is non-authoritative and neither is the current
MMR size. Note that this doesn't show the commitment *never*
existed, just that it's not an unspent commitment right now. The
guarantee of never-existence can't be made in the presence of
pruning nodes without breaking consensus.