Implement the sumtree archive receive logic
Gets the sumtree archive data stream from the network and write
it to a file. Unzip the file, place it at the right spot and
reconstruct the sumtree data structure, rewinding where to the
right spot.

Still 2 important missing pieces: validation, testing.

Note that the protocol is still very minimal and more checks
will be needed there as well.