Simplify sumtree backend buffering logic
The backend for a sumtree has to implement some in-memory
buffering logic to provide a commit/rollback interface. The
backend itself is an aggregate of 3 underlying storages
(an append only file, a remove log and a skip list). The buffering
was previously implemented both by the backend and some of the
underlying storages. Now pushing back all buffering logic to the
storages to keep the backend simpler.

This also allows for better reuse of underlying storages (notably
the append only file should get reused for kernels).