Fast sync handoff to body sync
Once the fast-sync state is fully setup, get bacj in body sync
mode to get the full bodies of the last blocks we're missing.

Also added fees to sumtree sum validation.

To fix: consuming directly from the TCPStream reader to get the
sumtree archive seems to hose the reader, so we don't get
anything after that. Time for another tokio and futures hair
pulling session.