Cleanup fast sync and few related bug fixes
Various cleanup. Consolidated fast sync and full sync into
a single function as they're very similar. Proper conditions to
trigger a sumtree request and some checks on receiving it.

Still needs more tests (for a simple reorg and when fast sync
needs to re-trigger for example). Also the case of a valid reorg
beyond the horizon isn't handled yet.