Prevent collision of headers sync attempts
Fixes #1357. It introduces a lock which prevents parallel handling of headers batches.
It also check if we process previous batch before sending a headers request. I simulated
the issue and tested the fix later on by setting header sync interval to 1 second. If we can
not process the current batch a new one is processed in parallel which makes the node even slower and
we start accumulating the debt until we basically stuck.