EC terminology edit proposal.
The original states "v1*H' is a value times an elliptic curve 'H'. Surely H should be a POINT on a elliptic curve shouldn't it? I've made an edit to line 95 to illustrate. If my edit is correct then a few more edits would be required further down the file.