Pin tokio and hyper versions
This is a tmp fix for  #2105, the current version of tokio doesn't allow
to call `block_on` on Runtime inside another Runtime. In our case we do
it to call API client from API handler. At the moment we switch to old
version of tokio and will implement a proper fix a bit later