Fix state sync for different partitions
Replacement of the txhahset does a rename to move the content of
the extracted folder. However one can't use rename across
different partitions (my case, `/tmp` has its own). In case of
failure, this re-extracts the zip to the destination.

Note that I originally tried a copy and delete, but recursively
copying the directory would either add a fair amount of code or
a dedicated dependency. This didn't seem worth (being somewhat of
a degenerate failure case).