Do not wrap Server into Arc
It is needed for p2p thread management, currently there is no point
where we could store thread handles and join themm because thread::join
consume the caller, which is impossible in case of Arc.
Also we don't need to check running flag constantly and it's easier to
reason about the state.