Introduce HTTP Client and proxy support
This PR adds support of HTTP proxy in Grin API client, which is useful
for some network setups and for Tor support. So now in addition to API
secret which we pass to functions like get/post we need to pass at
least one additional parameter. Also we instantiate HTTP clinet
connection objects each time we call such functions. To make
configuration management simpler and enable reuse of HTTP client a
struct GrinClient is introduced.