Represent Stream as a trait
We are interested only in behaviour of Stream, so it sounds right to
model it as a trait. It allows also reuse implementation of existing
traits (Read, Write) as is without wrapping. It simplifies client code
because it can work directly with TcpStream and I2pStream.

At the same time it brings some complexity - one more type parameter for
some functions and some boilerplate code to implement trait for
references.