Refactor needs_syncing
We return false in some "error" cases, eg when we don't have peers. It may trigger chain compaction for no reason, because we treat sync -> no sync state change as successful completion of syncing. So we should not return false in this case, we may just get a network error in the middle of the sync.
Also small code simplification